populate_task/requirements.txt
flask
psycopg2-binary
requests
musicbrainzngs
typing_extensions
